chore: (#544) OutingExceptions 통일
