--- dms-core/src/main/kotlin/team/aliens/dms/domain/outing/exception/OutingApplicationNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.outing.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.outing.exception.error.OutingErrorCode
-
-object OutingApplicationNotFoundException : DmsException(
-    OutingErrorCode.OUTING_APPLICATION_NOT_FOUND
-)
