--- dms-core/src/main/kotlin/team/aliens/dms/domain/outing/exception/OutingAvailableTimeMismatchException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.outing.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.outing.exception.error.OutingErrorCode
-
-object OutingAvailableTimeMismatchException :DmsException(
-    OutingErrorCode.OUTING_AVAILABLE_TIME_MISMATCH
-)
