+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/exception/OutingErrorCode.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.domain.outing.exception.error
+package team.aliens.dms.domain.outing.exception
 
 import team.aliens.dms.common.error.ErrorProperty
 import team.aliens.dms.common.error.ErrorStatus
@@ -10,8 +10,10 @@ enum class OutingErrorCode(
 ) : ErrorProperty {
 
     OUTING_AVAILABLE_TIME_MISMATCH(ErrorStatus.FORBIDDEN, "Outing Available Time Mismatch", 1),
-    OUTING_APPLICATION_EXISTS(ErrorStatus.CONFLICT, "Outing Application Exists", 1),
-    OUTING_APPLICATION_NOT_FOUND(ErrorStatus.NOT_FOUND, "Outing Not Found", 1)
+
+    OUTING_APPLICATION_NOT_FOUND(ErrorStatus.NOT_FOUND, "Outing Not Found", 1),
+
+    OUTING_APPLICATION_ALREADY_EXISTS(ErrorStatus.CONFLICT, "Outing Application Already Exists", 1),
     ;
 
     override fun status(): Int = status
